CB-8397 Add support to 'windows' for showing the Windows Phone splashscreen
